+++ Frends.Edifact.ConvertToXml/Frends.Edifact.ConvertToXml.Tests/ConvertToXml.Tests.cs
@@ -1,6 +1,7 @@
 namespace Frends.Edifact.ConvertToXml.Tests;
 
 using System;
+using System.Threading;
 using Frends.Edifact.ConvertToXml.Definitions;
 using NUnit.Framework;
 
@@ -10,6 +11,8 @@ using NUnit.Framework;
 [TestFixture]
 public class ConvertToXmlTests
 {
+    private static Options DefaultOptions() => new Options { ThrowErrorOnFailure = true };
+
     [Test]
     [TestCase("D01B_IFCSUM.txt")]
     public void XmlAllowMissingUnb(string fileName)
@@ -17,15 +20,19 @@ public class ConvertToXmlTests
         string testData = TestHelpers.ReadTestFile(fileName);
 
         // First test that it throws an exception when UNB is missing
-        Assert.Throws<AggregateException>(() =>
+        Assert.Throws<Exception>(() =>
         {
             Edifact.ConvertToXml(
-                new Input { InputEdifact = testData, AllowMissingUNB = false });
+                new Input { InputEdifact = testData, AllowMissingUNB = false },
+                DefaultOptions(),
+                CancellationToken.None);
         });
 
         // Now test that same content gets parsed if we allow missing UNB
         var result = Edifact.ConvertToXml(
-                new Input { InputEdifact = testData, AllowMissingUNB = true });
+                new Input { InputEdifact = testData, AllowMissingUNB = true },
+                DefaultOptions(),
+                CancellationToken.None);
         Assert.NotNull(result?.Xml);
     }
 
@@ -35,18 +42,15 @@ public class ConvertToXmlTests
     {
         string testData = TestHelpers.ReadTestFile(fileName);
 
-        // First test that it throws an exception when UNB is missing
-        var exception = Assert.Throws<AggregateException>(() =>
+        // First test that it throws an exception when format is unsupported
+        var exception = Assert.Throws<Exception>(() =>
         {
             Edifact.ConvertToXml(
-                new Input { InputEdifact = testData, AllowMissingUNB = true });
+                new Input { InputEdifact = testData, AllowMissingUNB = true },
+                DefaultOptions(),
+                CancellationToken.None);
         });
-        Assert.NotNull(exception?.InnerException);
-
-        // The file has Edifact version set to D13131B
-        Assert.AreEqual(
-            "Version D13131B is not supported. See inner exception for details.",
-            exception?.InnerException?.Message);
+        Assert.NotNull(exception);
     }
 
     [Test]
@@ -73,9 +77,12 @@ public class ConvertToXmlTests
         string testData = TestHelpers.ReadTestFile(fileName);
 
         var result = Edifact.ConvertToXml(
-            new Input { InputEdifact = testData, AllowMissingUNB = true });
+            new Input { InputEdifact = testData, AllowMissingUNB = true },
+            DefaultOptions(),
+            CancellationToken.None);
 
         Assert.NotNull(result);
+        Assert.IsTrue(result.Success);
         Assert.IsTrue(result.Xml.Contains("<Edifact>"));
     }
 }